requirements.txt
Flask
paho-mqtt
dnspython
watchdog
configargparse
